==> preprocessing/step4.ipynb <==
# Sports14 Text/Image Feature Extraction


import os
import numpy as np
import pandas as pd

os.chdir('/nfs/zixuan/MMRec/Amazon_Elec_Dataset')
os.getcwd()

## Load text data

i_id, desc_str = 'itemID', 'description'

file_path = './'
file_name = 'meta-elec.csv'

meta_file = os.path.join(file_path, file_name)

df = pd.read_csv(meta_file)
df.sort_values(by=[i_id], inplace=True)

print('data loaded!')
print(f'shape: {df.shape}')

df[:3]


# sentences: title + brand + category + description | All have title + description

title_na_df = df[df['title'].isnull()]
print(title_na_df.shape)

desc_na_df = df[df['description'].isnull()]
print(desc_na_df.shape)

na_df = df[df['description'].isnull() & df['title'].isnull()]
print(na_df.shape)

na3_df = df[df['description'].isnull() & df['title'].isnull() & df['brand'].isnull()]
print(na3_df.shape)

na4_df = df[df['description'].isnull() & df['title'].isnull() & df['brand'].isnull() & df['categories'].isnull()]
print(na4_df.shape)


df[desc_str] = df[desc_str].fillna(" ")
df['title'] = df['title'].fillna(" ")
df['brand'] = df['brand'].fillna(" ")
df['categories'] = df['categories'].fillna(" ")


# item_id
item_id = df[i_id].tolist()
print(len(item_id))

sentences = []
for i, row in df.iterrows():
    sen = row['title'] + ' ' + row['brand'] + ' '
    cates = eval(row['categories'])
    if isinstance(cates, list):
        for c in cates[0]:
            sen = sen + c + ' '
    sen += row[desc_str]
    sen = sen.replace('\n', ' ')

    sentences.append(sen)

# sentences[:10]
print(type(sentences))
print(len(sentences))
print(sentences[:10])

#make a dict item_id & text
dict_iid_text = dict(zip(item_id,sentences))
print(dict_iid_text[0])

#print for llama
print(dict_iid_text[24305])

import json
jsons = json.dumps(dict_iid_text)
with open("raw_text.json",'w') as f:
    f.write(jsons)
print('raw text json done')
