==> preprocessing/step6.ipynb <==
import numpy as np

# load numpy array from .npy file
feature_array = np.load('/nfs/zixuan/MMRec/data/clothing/image_feat_old.npy')

# create index list of item ids to be deleted
# missing_ids = [126, 328, 735, 836, 837, 839, 840, 841, 843, 1027, 1252, 1267, 1908, 1926, 2124, 2740, 3144, 3737, 5143, 5849, 6282, 6714, 7043, 7207, 7826, 7864, 8034, 8136, 8528, 9085, 9315, 9557, 9989, 10307, 10326, 10340, 10375, 10381, 10720, 10754, 10952, 10955, 11053, 11750, 11765, 12251, 12390, 12698, 12713, 13340, 13406, 13433, 13553, 13819, 13863, 13910, 14486, 14643, 14999, 15329, 15579, 16057, 16734, 17010, 17159, 17632, 17633, 17635, 17813, 17872]
# missing_ids = [284, 600, 967, 1255, 1282, 2046, 2480, 3790, 4123, 4365, 5209, 5303, 5328]
# missing_ids = [153, 1690, 3469, 3550, 5409, 5636, 6106, 8391, 8883, 9071, 9722, 10340, 10395, 10397, 10500, 10777, 10804, 10873, 11392, 11491, 11567, 11572, 11733, 11857, 12322, 20105, 20201, 21267, 22741, 26863, 28383, 28460, 28826, 28852, 29381, 29920, 30019, 30360, 32099, 32445, 32607, 33410, 34637, 34762, 35176, 37636, 41297, 42501, 43225, 43553, 45300, 45346, 48589, 48984, 54719, 55331, 55985, 60137, 61787, 61847, 62576, 62688]
missing_ids = [42, 43, 112, 169, 172, 176, 180, 193, 216, 241, 273, 297, 423, 432, 444, 465, 483, 518, 571, 572, 575, 576, 577, 579, 580, 593, 596, 606, 647, 648, 719, 723, 728, 747, 759, 760, 767, 913, 936, 940, 955, 1187, 1226, 1248, 1266, 1403, 1416, 1469, 1472, 1483, 1494, 1523, 1586, 1622, 1660, 1669, 1702, 1721, 1737, 1741, 1824, 1901, 1903, 2029, 2037, 2077, 2078, 2128, 2142, 2144, 2148, 2153, 2168, 2218, 2269, 2273, 2281, 2380, 2393, 2461, 2492, 2507, 2508, 2509, 2556, 2579, 2580, 2628, 2737, 2755, 2761, 2771, 2810, 2826, 2827, 2848, 2859, 2892, 2926, 3048, 3075, 3113, 3114, 3123, 3174, 3263, 3291, 3318, 3324, 3335, 3347, 3386, 3387, 3390, 3393, 3394, 3395, 3412, 3426, 3429, 3430, 3433, 3547, 3553, 3581, 3604, 3809, 3834, 3938, 4041, 4235, 4269, 4363, 4412, 4561, 4564, 4600, 4663, 4719, 4720, 4727, 4735, 4747, 4795, 4822, 4823, 4827, 4934, 4938, 4956, 4970, 5008, 5050, 5160, 5218, 5225, 5261, 5284, 5297, 5354, 5374, 5491, 5505, 5597, 5687, 5752, 5821, 5869, 5959, 6024, 6027, 6146, 6394, 6465, 6520, 6607, 6629, 6633, 6634, 6635, 6636, 6698, 6699, 6763, 6771, 6774, 6791, 6911, 6937, 6950, 6959, 6969, 7016, 7043, 7080, 7086, 7087, 7088, 7089, 7111, 7132, 7175, 7177, 7202, 7244, 7266, 7284, 7285, 7286, 7518, 7644, 7714, 7780, 7805, 7870, 7881, 7938, 7953, 7955, 8042, 8279, 8281, 8303, 8360, 8516, 8582, 8624, 8657, 8715, 8775, 8839, 8840, 8867, 9040, 9205, 9206, 9232, 9280, 9317, 9540, 9714, 9784, 9984, 9996, 10026, 10030, 10033, 10076, 10106, 10179, 10238, 10241, 10278, 10340, 10497, 10787, 10788, 10907, 11312, 11333, 11334, 11464, 11528, 11532, 11538, 11578, 11613, 11629, 11639, 11666, 11775, 11862, 12078, 12123, 12187, 12354, 12654, 12656, 12753, 12997, 13067, 13182, 13201, 13203, 13246, 13563, 13576, 13684, 13867, 13875, 13892, 13903, 14019, 14103, 14259, 14381, 14535, 14799, 14806, 14833, 14927, 15105, 15302, 15344, 15353, 15354, 15387, 15527, 15528, 15747, 15834, 15918, 15927, 15958, 15960, 15963, 15965, 15978, 15984, 16032, 16226, 16308, 16312, 16335, 16345, 16351, 16467, 16480, 16489, 16496, 16507, 16534, 16611, 16641, 16650, 16786, 16833, 16834, 17086, 17165, 17635, 17679, 17691, 17743, 18162, 18360, 18404, 18426, 18507, 18548, 18685, 18809, 18888, 18934, 18937, 19005, 19008, 19096, 19132, 19139, 19141, 19152, 19161, 19174, 19176, 19312, 19454, 19512, 19612, 19665, 19761, 19859, 19872, 19873, 19874, 19880, 19896, 19897, 19917, 19919, 19921, 19931, 20075, 20096, 20197, 20381, 20445, 20446, 20451, 20468, 20471, 20492, 20509, 20544, 20561, 20665, 20726, 20727, 20729, 20730, 20740, 20746, 20755, 20756, 20795, 20796, 20797, 20808, 20914, 20942, 20969, 21091, 21107, 21117, 21145, 21192, 21200, 21201, 21205, 21241, 21279, 21296, 21324, 21335, 21359, 21369, 21370, 21386, 21408, 21409, 21410, 21411, 21425, 21441, 21442, 21445, 21447, 21449, 21450, 21453, 21454, 21486, 21494, 21528, 21565, 21576, 21591, 21593, 21595, 21640, 21662, 21663, 21664, 21665, 21666, 21732, 21771, 21791, 21793, 21807, 21808, 21809, 21833, 21834, 21835, 21851, 21933, 21950, 21961, 21980, 21993, 21994, 21995, 22020, 22029, 22055, 22067, 22106, 22107, 22118, 22119, 22120, 22150, 22151, 22166, 22174, 22175, 22176, 22184, 22194, 22198, 22203, 22229, 22233, 22246, 22252, 22262, 22294, 22314, 22315, 22316, 22317, 22318, 22325, 22326, 22332, 22340, 22355, 22374, 22376, 22377, 22378, 22397, 22402, 22492, 22500, 22506, 22507, 22508, 22509, 22510, 22552, 22573, 22599, 22638, 22642, 22686, 22688, 22715, 22762, 22763, 22873, 22874, 22951]

# delete rows of feature array based on index list
feature_array = np.delete(feature_array, missing_ids, axis=0)

# print updated shape of feature array
print(feature_array.shape)

np.save('/nfs/zixuan/MMRec/data/clothing/image_feat_clothing.npy', feature_array)



import numpy as np

# load numpy array from .npy file
feature_array = np.load('/nfs/zixuan/MMRec/data/clothing/text_feat_old.npy')

# create index list of item ids to be deleted
# missing_ids = [126, 328, 735, 836, 837, 839, 840, 841, 843, 1027, 1252, 1267, 1908, 1926, 2124, 2740, 3144, 3737, 5143, 5849, 6282, 6714, 7043, 7207, 7826, 7864, 8034, 8136, 8528, 9085, 9315, 9557, 9989, 10307, 10326, 10340, 10375, 10381, 10720, 10754, 10952, 10955, 11053, 11750, 11765, 12251, 12390, 12698, 12713, 13340, 13406, 13433, 13553, 13819, 13863, 13910, 14486, 14643, 14999, 15329, 15579, 16057, 16734, 17010, 17159, 17632, 17633, 17635, 17813, 17872]
# missing_ids = [284, 600, 967, 1255, 1282, 2046, 2480, 3790, 4123, 4365, 5209, 5303, 5328]
# missing_ids = [153, 1690, 3469, 3550, 5409, 5636, 6106, 8391, 8883, 9071, 9722, 10340, 10395, 10397, 10500, 10777, 10804, 10873, 11392, 11491, 11567, 11572, 11733, 11857, 12322, 20105, 20201, 21267, 22741, 26863, 28383, 28460, 28826, 28852, 29381, 29920, 30019, 30360, 32099, 32445, 32607, 33410, 34637, 34762, 35176, 37636, 41297, 42501, 43225, 43553, 45300, 45346, 48589, 48984, 54719, 55331, 55985, 60137, 61787, 61847, 62576, 62688]
missing_ids = [42, 43, 112, 169, 172, 176, 180, 193, 216, 241, 273, 297, 423, 432, 444, 465, 483, 518, 571, 572, 575, 576, 577, 579, 580, 593, 596, 606, 647, 648, 719, 723, 728, 747, 759, 760, 767, 913, 936, 940, 955, 1187, 1226, 1248, 1266, 1403, 1416, 1469, 1472, 1483, 1494, 1523, 1586, 1622, 1660, 1669, 1702, 1721, 1737, 1741, 1824, 1901, 1903, 2029, 2037, 2077, 2078, 2128, 2142, 2144, 2148, 2153, 2168, 2218, 2269, 2273, 2281, 2380, 2393, 2461, 2492, 2507, 2508, 2509, 2556, 2579, 2580, 2628, 2737, 2755, 2761, 2771, 2810, 2826, 2827, 2848, 2859, 2892, 2926, 3048, 3075, 3113, 3114, 3123, 3174, 3263, 3291, 3318, 3324, 3335, 3347, 3386, 3387, 3390, 3393, 3394, 3395, 3412, 3426, 3429, 3430, 3433, 3547, 3553, 3581, 3604, 3809, 3834, 3938, 4041, 4235, 4269, 4363, 4412, 4561, 4564, 4600, 4663, 4719, 4720, 4727, 4735, 4747, 4795, 4822, 4823, 4827, 4934, 4938, 4956, 4970, 5008, 5050, 5160, 5218, 5225, 5261, 5284, 5297, 5354, 5374, 5491, 5505, 5597, 5687, 5752, 5821, 5869, 5959, 6024, 6027, 6146, 6394, 6465, 6520, 6607, 6629, 6633, 6634, 6635, 6636, 6698, 6699, 6763, 6771, 6774, 6791, 6911, 6937, 6950, 6959, 6969, 7016, 7043, 7080, 7086, 7087, 7088, 7089, 7111, 7132, 7175, 7177, 7202, 7244, 7266, 7284, 7285, 7286, 7518, 7644, 7714, 7780, 7805, 7870, 7881, 7938, 7953, 7955, 8042, 8279, 8281, 8303, 8360, 8516, 8582, 8624, 8657, 8715, 8775, 8839, 8840, 8867, 9040, 9205, 9206, 9232, 9280, 9317, 9540, 9714, 9784, 9984, 9996, 10026, 10030, 10033, 10076, 10106, 10179, 10238, 10241, 10278, 10340, 10497, 10787, 10788, 10907, 11312, 11333, 11334, 11464, 11528, 11532, 11538, 11578, 11613, 11629, 11639, 11666, 11775, 11862, 12078, 12123, 12187, 12354, 12654, 12656, 12753, 12997, 13067, 13182, 13201, 13203, 13246, 13563, 13576, 13684, 13867, 13875, 13892, 13903, 14019, 14103, 14259, 14381, 14535, 14799, 14806, 14833, 14927, 15105, 15302, 15344, 15353, 15354, 15387, 15527, 15528, 15747, 15834, 15918, 15927, 15958, 15960, 15963, 15965, 15978, 15984, 16032, 16226, 16308, 16312, 16335, 16345, 16351, 16467, 16480, 16489, 16496, 16507, 16534, 16611, 16641, 16650, 16786, 16833, 16834, 17086, 17165, 17635, 17679, 17691, 17743, 18162, 18360, 18404, 18426, 18507, 18548, 18685, 18809, 18888, 18934, 18937, 19005, 19008, 19096, 19132, 19139, 19141, 19152, 19161, 19174, 19176, 19312, 19454, 19512, 19612, 19665, 19761, 19859, 19872, 19873, 19874, 19880, 19896, 19897, 19917, 19919, 19921, 19931, 20075, 20096, 20197, 20381, 20445, 20446, 20451, 20468, 20471, 20492, 20509, 20544, 20561, 20665, 20726, 20727, 20729, 20730, 20740, 20746, 20755, 20756, 20795, 20796, 20797, 20808, 20914, 20942, 20969, 21091, 21107, 21117, 21145, 21192, 21200, 21201, 21205, 21241, 21279, 21296, 21324, 21335, 21359, 21369, 21370, 21386, 21408, 21409, 21410, 21411, 21425, 21441, 21442, 21445, 21447, 21449, 21450, 21453, 21454, 21486, 21494, 21528, 21565, 21576, 21591, 21593, 21595, 21640, 21662, 21663, 21664, 21665, 21666, 21732, 21771, 21791, 21793, 21807, 21808, 21809, 21833, 21834, 21835, 21851, 21933, 21950, 21961, 21980, 21993, 21994, 21995, 22020, 22029, 22055, 22067, 22106, 22107, 22118, 22119, 22120, 22150, 22151, 22166, 22174, 22175, 22176, 22184, 22194, 22198, 22203, 22229, 22233, 22246, 22252, 22262, 22294, 22314, 22315, 22316, 22317, 22318, 22325, 22326, 22332, 22340, 22355, 22374, 22376, 22377, 22378, 22397, 22402, 22492, 22500, 22506, 22507, 22508, 22509, 22510, 22552, 22573, 22599, 22638, 22642, 22686, 22688, 22715, 22762, 22763, 22873, 22874, 22951]

# delete rows of feature array based on index list
feature_array = np.delete(feature_array, missing_ids, axis=0)

# print updated shape of feature array
print(feature_array.shape)

np.save('/nfs/zixuan/MMRec/data/clothing/text_feat_clothing.npy', feature_array)
